Добавлен вывод состава корзины с подсчетом количества товаров и суммы стоимости товаров в корзине

+++ store_app/views.py
@@ -33,6 +33,27 @@ class CartView(APIView):
     """ Обрабатывает операции с корзиной: добавление, изменение, удаление товаров """
     permission_classes = [IsAuthenticated]
 
+    def get(self, request: Request) -> Response:
+        """ Выводит состав корзины с подсчетом количества товаров и суммы стоимости товаров в корзине """
+        cart, created = Cart.objects.get_or_create(user=request.user)
+
+        cart_product = CartProduct.objects.filter(cart=cart).select_related("product")
+        if not cart_product.exists():
+            return Response({
+                "products": [],
+                "total_quantity": 0,
+                "total_cost": 0
+            })
+
+        products_serializer = CartProductSerializer(cart_product, many=True)
+        total_quantity = sum(product.quantity for product in cart_product)
+        total_cost = sum(product.product.price for product in cart_product)
+        return Response({
+            "products": products_serializer.data,
+            "total_quantity": total_quantity,
+            "total_cost": float(total_cost)
+        })
+
     def post(self, request: Request) -> Response:
         """ Добавляет продукт в корзину """
         serializer = AddToCartSerializer(data=request.data, context={"request": request})
@@ -62,17 +83,32 @@ class CartView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def delete(self, request: Request) -> Response:
-        """ Удаляет продукт из корзины """
+        """ Удаляет продукт из корзины или очищает корзину полностью"""
+        clear_all = request.query_params.get("clear", False)
         product_id = request.query_params.get("product_id")
         product_id = int(product_id)
 
-        try:
-            cart = Cart.objects.get(user=request.user)
-            product_cart = CartProduct.objects.get(cart=cart, product_id=product_id)
-            product_cart.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except CartProduct.DoesNotExist:
+        cart = Cart.objects.get(user=request.user)
+
+        if clear_all:
+            CartProduct.objects.filter(cart=cart).delete()
             return Response(
-                {"error": "Продукт не найден в корзине"},
-                status=status.HTTP_404_NOT_FOUND
+                {"message": "Корзина очищена"},
+                status=status.HTTP_204_NO_CONTENT
+            )
+
+        if product_id:
+            try:
+                product_cart = CartProduct.objects.get(cart=cart, product_id=product_id)
+                product_cart.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            except CartProduct.DoesNotExist:
+                return Response(
+                    {"error": "Продукт не найден в корзине"},
+                    status=status.HTTP_404_NOT_FOUND
+                )
+        else:
+            return Response(
+                {"error": "Не указан product_id или параметр clear"},
+                status=status.HTTP_400_BAD_REQUEST
             )
